requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-openai
langchain-community
openai
python-dotenv
faiss-cpu
html2text